requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==3.2.13
django-crispy-forms==1.14.0
filelock==3.6.0
mssql-django==1.0rc1
django-betterforms==1.2
Pillow==9.1.0
platformdirs==2.5.2
pyodbc==4.0.32
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
python-dotenv
whitenoise
django-cors-headers
regex==2022.3.2
times==0.7
djangorestframework==3.13.1